requirements.txt
llama-index
pypdf
beautifulsoup4
requests
psycopg2-binary
python-dotenv
fastapi
uvicorn
sentence-transformers
torch
transformers
llama-index-readers-web
llama-index-vector-stores-postgres
llama-index[huggingface]
llama-index-embeddings-huggingface
asyncpg